back/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
boto3==1.26.45
python-dotenv==0.21.0
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
stripe
flasgger==0.9.5
flask-cors